kafka_to_pg/requirements.txt
bytewax==0.11.1
pymongo==4.2.0
kafka-python==2.0.2